refactor(idl-decode): name the IDL source modules uniformly
